backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.66.3
pydantic==2.10.6
python-multipart==0.0.20 